requirements.txt
asgiref==3.8.1
dj-database-url==3.0.0
Django==4.2.7
django-cors-headers==4.7.0
django-extensions==4.1
django-filter==25.1
django-ratelimit==4.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
gunicorn==23.0.0
mysqlclient==2.2.7
packaging==25.0
pillow==11.2.1
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-decouple==3.8
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2025.2
whitenoise==6.9.0
